extract interface from ConsoleInputHandler
