+++ src/main/java/quentin/UI/InputHandler.java
@@ -1,0 +1,10 @@
+package quentin.UI;
+
+import quentin.UI.console.ConsoleInputHandler;
+
+import java.util.InputMismatchException;
+
+public interface InputHandler {
+    int getInteger() throws InputMismatchException;
+    boolean askPie() throws InputMismatchException;
+}
