+++ src/main/java/quentin/UI/console/ConsoleStoneRepresentation.java
@@ -1,4 +1,4 @@
-package quentin.UI;
+package quentin.UI.console;
 
 import quentin.core.Stone;
